feat(llc): common hooks custom fees new send flow
