+++ apps/ledger-live-desktop/src/mvvm/features/Send/screens/CustomFees/components/CustomFeesScreenInner.tsx
@@ -7,7 +7,7 @@ import {
   getMainAccount,
 } from "@ledgerhq/ledger-wallet-framework/account/helpers";
 import { sendFeatures } from "@ledgerhq/live-common/bridge/descriptor/send/features";
-import { useStableGasOptions } from "../hooks/useStableGasOptions";
+import { useStableGasOptions } from "@ledgerhq/live-common/flows/send/customFees/hooks/useStableGasOptions";
 import { isEvmTransaction } from "../../../utils/isEvmTransaction";
 import { CustomFeesScreenInnerBase } from "./CustomFeesScreenInnerBase";
 import { CustomFeesScreenInnerWithAmountPlugins } from "./CustomFeesScreenInnerWithAmountPlugins";
