+++ apps/ledger-live-desktop/src/mvvm/features/Send/screens/CustomFees/hooks/__tests__/useStableGasOptions.test.ts
@@ -2,7 +2,7 @@
 import { BigNumber } from "bignumber.js";
 import { renderHook } from "tests/testSetup";
 import type { Transaction } from "@ledgerhq/live-common/generated/types";
-import { useStableGasOptions } from "../useStableGasOptions";
+import { useStableGasOptions } from "@ledgerhq/live-common/flows/send/customFees/hooks/useStableGasOptions";
 
 function createEvmTransaction(params: { gasOptions?: Record<string, unknown> }): Transaction {
   return {
