+++ libs/ledger-live-common/.unimportedrc.json
@@ -647,6 +647,11 @@
     "src/flows/send/coinControl/hooks/useCoinControlScreenViewModelCore.ts",
     "src/flows/send/coinControl/index.ts",
     "src/flows/send/coinControl/utils/changeToReturn.ts",
+    "src/flows/send/customFees/hooks/useBridgeFeeEstimation.ts",
+    "src/flows/send/customFees/hooks/useCustomFeeValidation.ts",
+    "src/flows/send/customFees/hooks/useCustomFeesViewModelCore.ts",
+    "src/flows/send/customFees/hooks/useStableGasOptions.ts",
+    "src/flows/send/customFees/utils/customFeeUtils.ts",
     "src/flows/send/hooks/useSendFlowAccount.ts",
     "src/flows/send/hooks/useSendFlowAmountReviewCore.ts",
     "src/flows/send/hooks/useSendFlowBusinessLogic.ts",
