+++ libs/ledger-live-common/src/flows/send/customFees/hooks/useCustomFeeValidation.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
+import { BigNumber } from "bignumber.js";
 import type { AccountLike } from "@ledgerhq/types-live";
-import type { Transaction, TransactionStatus } from "@ledgerhq/live-common/generated/types";
-import type { CustomFeeInputDescriptor } from "@ledgerhq/live-common/bridge/descriptor/types";
-import BigNumber from "bignumber.js";
+import type { Transaction, TransactionStatus } from "../../../../coin-modules/transaction-types";
+import type { CustomFeeInputDescriptor } from "../../../../bridge/descriptor/types";
 import {
   isValidNumberForInput,
   computeMinValue,
